move oc init into a separate component
